products/serializers/subcategory.py
from rest_framework import serializers
from products.models import SubCategory

class SubCategorySerializer(serializers.ModelSerializer):

    class Meta:
        model=SubCategory
        fields=["name","id","createdat"]